requirements.txt
requests
loguru
pycryptodome
fastapi[standard]